Prepare for npm
Since there is no needed docker anymore (npm env) dockerfiles were removed.
Also makefile is removed, lib directory is no more needed (all is root now)
Changed the route to the dist and sites directory to map to requirer.
easy-live-reload removed due to stopped working correctly
