+++ .gitignore
@@ -1,1 +1,6 @@
+node_modules
 .DS_Store
+*.log
+.env
+dist
+sites
