+++ config/index.js
@@ -1,0 +1,9 @@
+import dotenv from 'dotenv'
+import './logger'
+
+dotenv.config()
+const config = {}
+
+config.PORT = process.env.PORT
+
+global.config = config
