--- docker-compose.yml
@@ -1,25 +1,0 @@
-version: "3"
-services:
-  development:
-    build: ./lib/.
-    container_name: lego-landing
-    volumes:
-      - ./lib/.:/usr/app
-      - ../sites/src/.:/usr/app/sites/
-      - /usr/app/node_modules
-    environment:
-      - ENV=development
-      - PORT=80
-    ports:
-      - '8080:80'
-      - '35729:35729'
-    command: npm start
-  build:
-    build: ./lib/.
-    container_name: lego-builder
-    volumes:
-      - ./lib/.:/usr/app
-      - ../sites/src/.:/usr/app/sites/
-      - ../sites/dist/.:/usr/app/dist/
-      - /usr/app/node_modules
-    command: npm run build
