+++ index.js
@@ -1,0 +1,5 @@
+require('babel-register')
+const { Server, Builder } = require('./src')
+
+module.exports.Builder = Builder
+module.exports.Server = Server
