--- lib/.gitignore
@@ -1,6 +1,0 @@
-node_modules
-dist
-sites
-.DS_Store
-*.log
-.env
