--- lib/Dockerfile
@@ -1,9 +1,0 @@
-FROM node:10
-
-WORKDIR /usr/app
-
-COPY package.json .
-
-RUN npm install
-
-COPY . .
