--- lib/config/index.js
@@ -1,6 +1,0 @@
-import './logger'
-const config = {}
-
-config.PORT = process.env.PORT || 1337
-
-global.config = config
