--- lib/nodemon.json
@@ -1,4 +1,0 @@
-{
-  "ext": "js,json",
-  "ignore": ["./sites/"]
-}
