--- lib/src/main.js
@@ -1,2 +1,0 @@
-import '../config'
-import './server'
