+++ package.json
@@ -3,10 +3,6 @@
   "version": "1.0.0",
   "description": "Landing Maker",
   "main": "index.js",
-  "scripts": {
-    "start": "nodemon index.js",
-    "build": "node build.js"
-  },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/resuelve/lego-landing.git"
@@ -24,16 +20,17 @@
     "body-parser": "^1.18.3",
     "consolecolors": "^0.33.2-b",
     "copy-dir": "^0.3.0",
-    "easy-livereload": "^1.3.1",
+    "dotenv": "^6.0.0",
     "express": "^4.16.3",
     "human-elapsed": "^1.0.3",
     "memoizee": "^0.4.12",
     "mkdirp": "^0.5.1",
-    "nodemon": "^1.17.5",
     "pug": "^2.0.3",
     "request": "^2.87.0",
     "rimraf": "^2.6.2",
     "sass": "^1.5.0"
   },
-  "dependencies": {}
+  "dependencies": {
+    "livereload": "^0.7.0"
+  }
 }
