+++ src/index.js
@@ -1,0 +1,6 @@
+import '../config'
+const { Server } = require('./server')
+const { Builder } = require('./builder')
+
+module.exports.Builder = Builder
+module.exports.Server = Server
